feat: queue a batch of unique transaction builders at once (#10)
Adding this PoC that has as objective cover the following features:
- we should be able to enqueue multiple action addresses at the same
time
- an action address that's currently on the queue shouldn't be
enqueuable again

For this, we're modifying how we store the actions in the contract,
instead of doing `hash -> bytes`, we can do `hash -> struct` that holds
more information, for instance, an array of the action contracts, that
can be updated when the action is unqueued or executed, so that it keeps
track of the `queuedActions` mapping, such that it controlls the
`ActionAlreadyQueued` error.

This implementation does requires a bit more array handling, which may
be quite gas expensive and perhaps prone to errors, we shall need, in
order to complete the PoC, an integration test that's sensitive to
multiple actions being batched. Also, we shall re-write the other
methods (i.e. `queueApprovedAction`) to make use of this struct.

This PR is just for explaining a possible approach to the requested
features.

Closes BES-464.

---------

Co-authored-by: Lempire <61431140+lempire123@users.noreply.github.com>
Co-authored-by: 0xJabberwock <0xjabberwock@defi.sucks>